👷 commit_users

+++ Backend/modules/users.js
@@ -1,7 +1,5 @@
 const pool = require('../core/config_db');
 
-require('dotenv').config();
-
 const getUsers = async function () {
     try {
         const res = await pool.query('SELECT * FROM users');
@@ -12,13 +10,19 @@ const getUsers = async function () {
     }
 };
 
-const addUser = async function (user_id, name, phone, email, password_information, driver_id ) {
-
-    const driverCheckRes = await pool.query('SELECT * FROM drivers WHERE id = $1', [driver_id]);
-    if (driverCheckRes.rowCount === 0) {
-        throw new Error(`Driver with ID ${driver_id} does not exist`);
-    }
+const addUser = async function (user) {
+    const { user_id, name, phone, email, password_information, driver_id } = user;
     try {
+        const driverCheckRes = await pool.query('SELECT * FROM drivers WHERE id = $1', [driver_id]);
+        if (driverCheckRes.rowCount === 0) {
+            throw new Error(`Driver with ID ${driver_id} does not exist`);
+        }
+
+        const userCheckRes = await pool.query('SELECT * FROM users WHERE user_id = $1', [user_id]);
+        if (userCheckRes.rowCount > 0) {
+            throw new Error('User with ID already exists');
+        }
+
         const res = await pool.query(`
             INSERT INTO users (user_id, name, phone, email, password_information, driver_id)
             VALUES ($1, $2, $3, $4, $5, $6)
@@ -29,27 +33,29 @@ const addUser = async function (user_id, name, phone, email, password_informatio
         console.error('Error adding user:', error);
         throw error;
     }
-}
-const getUser = async function (user_id) {
+};
 
+const getUser = async function (user_id) {
     try {
-        const res = await pool.query(`SELECT * FROM users WHERE user_id = $1`, [user_id]);
+        const res = await pool.query('SELECT * FROM users WHERE user_id = $1', [user_id]);
         return res.rows[0];
     } catch (error) {
         console.error('Error getting user:', error);
         throw error;
-
     }
 };
 
-
 const deleteUser = async function (user_id) {
     try {
-        const res = await pool.query(`
-            DELETE FROM users
-            WHERE user_id = $1
-            RETURNING *
-        `, [user_id]);
+        const userCheckRes = await pool.query('SELECT * FROM users WHERE user_id = $1', [user_id]);
+        if (userCheckRes.rowCount === 0) {
+            throw new Error('User not found');
+        }
+
+        const res = await pool.query('DELETE FROM users WHERE user_id = $1 RETURNING *', [user_id]);
+        if (res.rowCount === 0) {
+            throw new Error('Error deleting user');
+        }
         return res.rows[0];
     } catch (error) {
         console.error('Error deleting user:', error);
@@ -60,12 +66,22 @@ const deleteUser = async function (user_id) {
 const updateUser = async function (reqBody) {
     const { user_id, name, phone, email, password_information, driver_id } = reqBody;
     try {
+        const userCheckRes = await pool.query('SELECT * FROM users WHERE user_id = $1', [user_id]);
+        console.log("????????????????");
+        console.log(userCheckRes);
+        if (userCheckRes.rowCount === 0) {
+            throw new Error('User does not exist');
+        }
+
         const res = await pool.query(`
-            UPDATE users 
+            UPDATE users
             SET name = $1, phone = $2, email = $3, password_information = $4, driver_id = $5
             WHERE user_id = $6
             RETURNING *
         `, [name, phone, email, password_information, driver_id, user_id]);
+        if (res.rowCount === 0) {
+            throw new Error('Error deleting user');
+        }
         return res.rows[0];
     } catch (error) {
         console.error('Error updating user:', error);
